adding user profile button & dropdown menu
